Support dev/test/prod database environments
This changes how the `~/.blusedb.yml` file needs to be structured.
Basically, it adds a top-level hierarchy on top of the already existing
hierarchy.  This new top-level hierarchy supports three keys: `dev`,
`test`, and `prod`.  Each top level key contains a standalone database
configuration for development, testing, and production, resp.  Each of
these configurations supports the use of "levels" to provide varying
degrees of access to the database (e.g. by specifying different
usernames and passwords for different levels of access).  The three
levels of access are: `admin`, `user`, and `reader`.  Parameters that
are common across all levels can be specified outside of a "level" (but
still under the environment.

The structure is now:

    dev:
      host: DEVHOST
      database: BLUSEDB_DEV
      admin:
        username: ADMIN_USER
        password: ADMIN_PASS
      user:
        username: USER_USER
        password: USER_PASS
      reader:
        username: READER_USER
        password: READER_PASS

    test:
      host: TESTHOST
      database: BLUSEDB_TEST
      admin:
        username: ADMIN_USER
        password: ADMIN_PASS
      user:
        username: USER_USER
        password: USER_PASS
      reader:
        username: READER_USER
        password: READER_PASS

    prod:
      host: PRODHOST
      database: BLUSEDB_PROD
      admin:
        username: ADMIN_USER
        password: ADMIN_PASS
      user:
        username: USER_USER
        password: USER_PASS
      reader:
        username: READER_USER
        password: READER_PASS

Note that fields shown in uppercase should be replaced with site
specific values.  It is possible (even common) for DEVHOST, TESTHOST,
and PRODHOST to all be on the same host.