Improve RawFile insert!() logic
It used to be "insert and select on insert failure".  Now it is "select
and insert on not found and select on insert failure".  This will
minimize but not eliminate failed inserts when multiple hosts are
inserting.  I think eliminating these failed inserts would require some
form of MySQL table locks, but it's probably not worth the effort.